Extract Key* classes to a separate file
